+++ cdx_core/Gemfile
@@ -41,6 +41,7 @@ gem 'nuntium_api', '~> 0.21'
 gem 'config'
 gem 'decent_exposure', '2.3.2'
 gem 'gon'
+gem 'rubyzip', '>= 1.0.0'
 
 group :development do
   gem 'letter_opener'
