+++ cdx_core/spec/dummy/config/environments/test.rb
@@ -31,6 +31,8 @@ Rails.application.configure do
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
+  config.action_mailer.default_url_options = { host: Settings.host }
+
   # Randomize the order test cases are executed.
   config.active_support.test_order = :random
