+++ cdx_core/spec/dummy/config/initializers/oj.rb
@@ -1,0 +1,7 @@
+Oj.default_options = {mode: :compat}
+
+class Object
+  def to_json_oj
+    Oj.dump self
+  end
+end
